backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
websockets
python-dotenv
pydantic
google-cloud-firestore
